third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-do-not-bind-dirty-sourcecode.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../debugger-test.js"></script>
<script src="../console-test.js"></script>
<script src="../isolated-filesystem-test.js"></script>
<script src="./persistence-test.js"></script>
<script src="./resources/foo.js"></script>
<script>

function test()
{
    Runtime.experiments.enableForTest('persistence2');
    var testMapping = BindingsTestRunner.initializeTestMapping();
    var fs = new BindingsTestRunner.TestFileSystem("file:///var/www");
    BindingsTestRunner.addFooJSFile(fs);
    fs.reportCreated(function() { });

    TestRunner.runTestSuite([
        function waitForUISourceCodes(next)
        {
            Promise.all([
                TestRunner.waitForUISourceCode("foo.js", Workspace.projectTypes.Network)
                    .then(uiSourceCode => uiSourceCode.setWorkingCopy("dirty.")),
                TestRunner.waitForUISourceCode("foo.js", Workspace.projectTypes.FileSystem)
            ]).then(next);
        },

        function addFileSystemMapping(next)
        {
            TestRunner.addSniffer(Persistence.Persistence.prototype, '_prevalidationFailedForTest', onPrevalidationFailed);
            testMapping.addBinding('foo.js');

            function onPrevalidationFailed(binding)
            {
                TestRunner.addResult("Failed to create binding: " + binding);
                next();
            }
        },
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Verify that dirty uiSourceCodes are not bound.</p>
</body>
</html>
